api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase-py
pydantic
